update sdk for 7.27, reaper_plugin_functions.h now up to date

+++ reaper-plugins/reaper_mp3/mpglib/layer2.cpp
@@ -78,14 +78,14 @@ int mpglib::II_step_one(unsigned int *bit_alloc,int *scale,struct frame *fr)
     bita = bit_alloc;
     if(stereo)
     {
-		  for (i=jsbound;i;i--,alloc1+=(size_t) (1<<step))
+		  for (i=jsbound;i;i--,alloc1+=(size_t)(int)(1<<step))
 		  {
 			  step=alloc1->bits;
 			  if (step<=0) return 0;
 			  *bita++ = (char) getbits(step);
 	      *bita++ = (char) getbits(step);
 		  }
-		  for (i=sblimit-jsbound;i;i--,alloc1+=(size_t) (1<<step))
+		  for (i=sblimit-jsbound;i;i--,alloc1+=(size_t)(int) (1<<step))
 		  {
 			  step=alloc1->bits;
 			  if (step<=0) return 0;
@@ -101,7 +101,7 @@ int mpglib::II_step_one(unsigned int *bit_alloc,int *scale,struct frame *fr)
 		}
 		else /* mono */
 		{
-		  for (i=sblimit;i;i--,alloc1+=(size_t) (1<<step))
+		  for (i=sblimit;i;i--,alloc1+=(size_t)(int)(1<<step))
 		  {
 			  step=alloc1->bits;
 			  if (step<=0) return 0;
@@ -155,7 +155,7 @@ int mpglib::II_step_two(unsigned int *bit_alloc,real fraction[2][4][SBLIMIT],int
     unsigned int *bita=bit_alloc;
     int d1,step;
 
-    for (i=0;i<jsbound;i++,alloc1+=(size_t) (1<<step))
+    for (i=0;i<jsbound;i++,alloc1+=(size_t)(int)(1<<step))
     {
       step = alloc1->bits;
       for (j=0;j<stereo;j++)
@@ -187,7 +187,7 @@ int mpglib::II_step_two(unsigned int *bit_alloc,real fraction[2][4][SBLIMIT],int
       }
     }
 
-    for (i=jsbound;i<sblimit;i++,alloc1+=(size_t) (1<<step))
+    for (i=jsbound;i<sblimit;i++,alloc1+=(size_t)(int)(1<<step))
     {
       step = alloc1->bits;
       bita++;	/* channel 1 and channel 2 bitalloc are the same */
